Moving customization to central package
